+++ tests/components/caldav/conftest.py
@@ -1,6 +1,6 @@
 """Test fixtures for caldav."""
 from collections.abc import Awaitable, Callable
-from unittest.mock import patch
+from unittest.mock import Mock, patch
 
 import pytest
 
@@ -17,6 +17,24 @@ UNIQUE_ID = "unique-id-1"
 def mock_platforms() -> list[Platform]:
     """Fixture to specify platforms to test."""
     return []
+
+
+@pytest.fixture(name="calendars")
+def mock_calendars() -> list[Mock]:
+    """Fixture to provide calendars returned by CalDAV client."""
+    return []
+
+
+@pytest.fixture(name="dav_client", autouse=True)
+def mock_dav_client(calendars: list[Mock]) -> Mock:
+    """Fixture to mock the DAVClient."""
+    with patch(
+        "homeassistant.components.caldav.calendar.caldav.DAVClient"
+    ) as mock_client:
+        mock_client.return_value.principal.return_value.calendars.return_value = (
+            calendars
+        )
+        yield mock_client
 
 
 @pytest.fixture(name="config_entry")
